完善导出功能

+++ src/main/java/cc/mrbird/febs/system/entity/Kehu.java
@@ -58,7 +58,7 @@ public class Kehu {
      */
     @ExcelIgnore
     @TableField("ssyh")
-    private Integer ssyh;
+    private Long ssyh;
 
     /**
      * 联系人电话
